Remove command-line login and isolate login logic
